moved crawling into processors
